backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]
# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database (Supabase PostgreSQL)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
supabase==2.3.4
asyncpg==0.29.0

# MARE AI/ML dependencies
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
scipy==1.11.1

# Feature engineering and preprocessing
category_encoders==2.6.1
imbalanced-learn==0.11.0

# Recommendation systems
surprise==1.1.3
implicit==0.7.2

# CAST Framework - Multilingual NLP and Translation
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99
sacremoses==0.1.1

# Language detection and processing
langdetect==1.0.9
polyglot==16.7.4

# Cultural and bias analysis
spacy==3.7.2
nltk==3.8.1

# Translation services integration (optional)
googletrans==4.0.0rc1

# Groq LLM for enhanced AI suggestions
groq==0.4.1

# Text processing and analysis
textblob==0.17.1
fuzzywuzzy==0.18.0
python-levenshtein==0.21.1

# Geographic processing (optional)
geopy==2.3.0
# geopandas==0.13.2  # Comment out for lighter installation

# Text processing for Indian languages (optional)
# polyglot==16.7.4  # Comment out if not using NLP features
# PyICU==2.11
# pycld2==0.41

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
loguru==0.7.2

# Optional: If you want to migrate to PostgreSQL later
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.23
